final commit for tonite lots of errors
