+++ package.json
@@ -9,7 +9,7 @@
     "client:build": "react-scripts build",
     "server:dev": "nodemon index.js",
     "db:build": "nodemon  db/seed.js",
-    "start:dev": "concurrently \"npm run client:build\" \"npm run client:dev\" \"npm run server:dev\"",
+    "start:dev": "concurrently \"npm run client:build\" \"npm run server:dev\" \"npm run client:dev\"",
     "start": "node index.js",
     "test": "jest --runInBand",
     "test:watch": "jest --runInBand --watchAll"
@@ -29,8 +29,8 @@
     "morgan": "^1.10.0",
     "pg": "^8.2.1",
     "react": "^16.14.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.3.4",
+    "react-dom": "^16.14.0",
+    "react-router-dom": "^6.14.1",
     "react-scripts": "^5.0.1"
   },
   "devDependencies": {
