+++ src/axios-services/index.js
@@ -1,4 +1,4 @@
-const BASE_URL = ``
+const BASE_URL = `https://localhost:3000/api`
 
 // this file holds your frontend network request adapters
 // think about each function as a service that provides data
@@ -7,50 +7,7 @@ const BASE_URL = ``
 // for example, if we need to display a list of users
 // we'd probably want to define a getUsers service like this:
 //========USER ENDPOINTS===========
-//Register
-export const register =async (user) =>{
-    try {
-      const response = await fetch(
-        `${BASE_URL}/users/register`,{
-          method: "POST",
-          headers:{
-            "Content-Type":"application/json"
-          },
-          body: JSON.stringify({
-            username: user.username,
-            password: user.password
-          })
-        }
-        )
-        const result = await response.json()
-        console.log("REGISTER RESULTS", result)
 
-        return result;
-    } catch(err) {
-      console.error(err)
-    }
-  }
-//Login
-export const login = async (user) => {
-      
-  try {
-    const response = await fetch(`${BASE_URL}/users/login`, {
-      method: "POST",
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-          username: user.username,
-          password: user.password
-      })
-    });
-    const result = await response.json();
-    console.log(result);
-    return result
-  } catch (err) {
-    console.error(err);
-  }
-}
 
 
 export async function getAPIHealth() {
