+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import {Routes, Route, useNavigate} from "react-router-dom"
-import {Register} from "./index"
+import {Registering} from "./index"
 // getAPIHealth is defined in our axios-services directory index.js
 // you can think of that directory as a collection of api adapters
 // where each adapter fetches specific info from our express server's /api route
@@ -10,12 +10,14 @@ import '../style/App.css';
 const App = () => {
   const [token, setToken] = useState("")
   const [isLoggedIn, setIsLoggedIn] = useState(false)
-
+  // const navigate = useNavigate();
+  
   function tokenCheck(){
     if(window.localStorage.getItem("token")){
       setToken(window.localStorage.getItem("token"))
     }
   }
+
 
   useEffect(() =>{
     tokenCheck();
@@ -30,18 +32,20 @@ const App = () => {
 
 
   return (
+    <>
+    <h1>OTTERS FROLICKING</h1>
     <div className="app-container">
-      <h1>OTTERS FROLICKING</h1>
-      <p></p>
       <Routes>
         <Route
-        path="/register"
-        element={<Register
-        setToken={setToken}
-        navigate={navigate}/>}
+        path='/Registering'
+        element={<Registering
+        
+        />}
         />
       </Routes>
+      
     </div>
+      </>
   );
 };
